+++ order-events-consumer/src/main/java/com/kodebytes/acasado/repository/OrderEventRepository.java
@@ -1,8 +1,8 @@
 package com.kodebytes.acasado.repository;
 
-import com.kodebytes.acasado.domain.OrderEvent;
+import com.kodebytes.acasado.entity.OrderEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderEventRepository extends JpaRepository<OrderEvent, Long> {
+public interface OrderEventRepository extends JpaRepository<OrderEvent, Integer> {
 }
