+++ order-events-consumer/src/main/java/com/kodebytes/acasado/repository/PhoneRepository.java
@@ -1,8 +1,8 @@
 package com.kodebytes.acasado.repository;
 
-import com.kodebytes.acasado.domain.Phone;
+import com.kodebytes.acasado.entity.Phone;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhoneRepository extends JpaRepository<Phone, Long> {
+public interface PhoneRepository extends JpaRepository<Phone, Integer> {
 }
